requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-nested-routers==0.94.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
pdfkit==1.0.0
pillow==11.1.0
platformdirs==4.3.7
psycopg==3.2.5
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
qrcode==8.1
reportlab==4.3.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zeep==4.3.1
